model cleanups
