+++ wasm-rpc-stubgen/src/model/mod.rs
@@ -1,2 +1,3 @@
-pub mod unknown_properties;
-pub mod wasm_rpc;
+pub mod app;
+pub mod app_raw;
+pub mod template;
